Made significant changes to the map class
added a constructor in the map class that constructs a map by extracting
the map info from a text file. also added a move function in the player
class that moves the player object in a direction as directed by an
integer (later to be identified with a keyboard event). Also made a new
Test class to test out the current progress of the project.