Fixed the test class
The test class works, with a default map, a player and a character on
the map.  All necessary changes were made to the character and the
player class for the test to function properly. Also many comments were
added throughout the project. The test class is now ready to implement
the key events. That's on you Japhet!